Fix vertical alignment of PALcolour output.
The output was off by a line after the fix to the initialisation of the
line pointers. This was the result of the "+ 2" in the output buffer
pointer setup -- the first output line should be at the top of the
buffer.

Checking alignment with Test Card G from Jason and the Argonauts showed
that the bottom line was also being cut off -- the Y loop had the wrong
upper bound. TCG is now vertically symmetrical.