Centre the decoded image.
For both PAL and NTSC, rather than adding extra pixels at the right when
adjusting the width, add them at both sides to keep the image within a
pixel of the centre.

Test Card G and the Video Essentials testcard aren't perfectly centred
with this, but they're now pretty close.