Implement 2D Transform PAL decoder.
The frequency-domain filter here is a direct translation of Jim
Easterbrook's implementation from pyctools-pal. It's only a 2D
implementation for now, but could be extended to 3D without too much
difficulty.

At present, it basically works, and shows a considerable reduction in
cross-colour artefacts over PALcolour (e.g. in the narrow street scene
on GGV1011). It produces quite a bit of dot patterning in luma, though.
The FFT size and default threshold came from pyctools-pal, and may
benefit from tuning for ld-decode's use. The window function is also
rather simplistic.

To use this, specify -f transform2d. This is implemented as a plugin to
PalDecoder/PalColour rather than a new Decoder class, because that
allows us to reuse PALcolour's chroma LPFs and demodulator. In
transform2d mode, PALcolour's luma filter output isn't used (I did try
filtering the Transform PAL luma through it, but the difference was
barely perceptible with the current settings).

transform2d mode is about half the speed of pal2d mode.

This adds a dependency on the FFTW library for FFT computations.