Make Comb's configuration work more like PalColour's.
Comb now uses VideoParameters directly, which avoids duplicating all the
fields and copying them around.

Comb::Configuration now has the default configuration parameters in when
constructed, so you don't need to create a Comb to get them.

I've changed the default for firstActiveLine from 43 to 40 to match the
first active line that the rest of the code uses for NTSC, but this
doens't make a difference between both ld-analyse and ld-chroma-decoder
set it explicitly anyway.

As a side-effect of this, Comb now uses the horizontal active region
values after setVideoParameters has adjusted them, so NTSC decodes will
use the full horizontal width of the padded frame, as PAL decodes
already do. (This is quite visible on the 3D map output because the
purple blobs now go all the way to the edges, rather than being
pillarboxed.)