Add Transform PAL 3D implementation.
The FFT structure is as described on Jim Easterbrook's page: each field
is expanded to a full frame, with black lines being inserted to maintain
the odd or even lines at the right spatial positions -- so the FFT sees
a sequence of images at 50 Hz, each of which has half the lines filled
in.

This is the higher-quality (but slower) choice of the two options that
the original authors considered for 3D structure; the other one could be
implemented as another TransformPal subclass in the future if needed.

The frequency-domain filter is the existing one for 2D derived from
pyctools-pal, with an additional loop to check for reflections in time.
(Note the XXX comment in there about the coordinate of the subcarrier on
the Z axis not being what I'd expected.)

The TransformPal configuration defaults remain the same; both modes work
in 2D and 3D, although from some initial testing "level" mode might be a
better default for 3D. ld-analyse can only use 2D filters at the moment,
so selecting Transform PAL there will get the equivalent of transform2d.