Fix ld-analyse crash on exit while freeing QCharts.
QChart::addAxis and QChart::addSeries take ownership of the pointers
they're given. However, in the dropout and SNR dialog code, the pointers
being used are to member variables of the dialog classes -- which caused
a crash when QChart's destructor tried to free the axes and series that
it thought it had ownership of.

Avoid this by explicitly taking ownership back in the dialogues'
destructors.