When no replacement line is found, don't correct at all.
Previously this corrected by using the previous equivalent line
(fieldLine - stepAmount) -- even if it was right at the top of the
image. Building with -fsanitize=address revealed that it was sometimes
reading outside the image bounds as a result.

However, this will only happen if findPotentialReplacementLines hasn't
found any candidate lines -- which means either we're outside the active
area, or the video is so bad that every other line has a dropout too.
Either way there's no point in blindly correcting from the previous
line.