Simplify the logic that finds the nearest replacement line.
Rather than deciding which is the nearest line in each field and then
choosing between those, put all the candidates in a vector and select
the nearest based on the *frame* line they'll end up on. This will also
be slightly more accurate than just looking at the field line numbers.