Initial support for subcarrier-locked sampling.
For 4fSC subcarrier-locked PAL, the TBC file contains four more samples
than usual over the course of 625 lines, so the two fields are
horizontally misaligned by two samples. This shifts the second field to
the left to compensate.

This is only an initial fix for now, since the same problem will apply
to ld-dropout-correct and the other tools -- it really needs dealing
with in the library.