When seeking forward isn't possible, try skipping.
When SourceVideo::getVideoField is reading video from an unseekable
source, like a pipe, trying to seek forward to a field later in the file
would fail -- which, for example, means you couldn't use
"ld-chroma-decoder -s" when input was from a pipe.

Make it try to read and discard data instead when seeking forward fails.
If reading fails or hits the end of the input stream, it'll report an
error as before.