Remove automatic chroma gain.
This is no longer necessary because ld-decode has MTF compensation and
produces reasonable chroma levels by default.

--chroma-gain now defaults to 1.0, and works for both PAL and NTSC.

As with the old PAL code, the NTSC YIQ to RGB conversion wasn't applying
an appropriate scaling factor to the IQ components. Poynton "Digital
Video and HDTV" 1st edition p514 says that the scaling factor is the
same as for Y -- i.e. when 7.5% setup is in use, the chroma range is
proportionately smaller.