Use the 3D rather than 2D result in ntsc3d.
After the removal of the OpenCV code in 6c7f74d663d ("Fix for
issue #484"), ntsc3d always chooses the 2D result, so it behaves the
same as ntsc2d. This makes it always use the 3D result instead, which
produces artefacts on motion, but much better quality than ntsc2d for
still images.