Implement a new ntsc3d adaptive filter.
3D decoding looks for candidates in the current frame and two fields in
either direction, using a similarity heuristic based on the output of
the existing 2D decoder.

The heuristic is a quality tradeoff: it can produce worse results than
the non-adaptive 3D decoder for absolutely still images (because of
spurious chroma returned by the 2D decoder, and content that's similar
enough to fool the 3D heuristic), but overall the results are much
better, particularly for animation and other material with repeated
fields/frames.