Make TbcSource have a "current frame".
Previously all the methods that got data from TbcSource took a frame
number as an argument, which meant they all had to load frame metadata
individually.

There's now a loadFrame method that you call when the current frame
changes (which only happens in one place in ld-analyse), and this caches
the metadata for the other methods to use directly.

The chapter search methods still take a frame number because they're
only searching the chapter map.