Always enable the NTSC equiband IQ filter.
Since it's always on in phase-compensating mode anyway, and it provides
a small improvement in non-phase-compensating mode, we may as well
remove the option to disable it entirely.

Narrowband mode as currently implemented makes a negligable difference
to the output, and Poynton says that narrowband encoding "has not been
practiced since about 1970", so remove that as well. (Poynton also notes
that in theory a narrowband decoder should use a different shape for the
I filter as well.)