Use a hash table to decode EFM symbols.
The existing code did a linear search across a 256-entry array. Since
this fits in L1 cache it wasn't as bad as it looked, but doing it with a
hash table instead reduces the number of memory reads by a factor of
100 (at most 2 reads per lookup).

The hash function here was designed by an exhaustive search using a
Python program.