Restructure Demodulator so it can be vectorised.
This works the same way as the existing version, but it reworks the
buffer code to avoid using % on the indexes in the loop. This means that
GCC can use vector instructions, and schedule the unrolled outer loop a
bit more effectively.

I've also replaced 320 etc. with the derivations of the constants.