Use double rather than qreal in remaining tools code.
This changes the metadata fields that were qreal to double, which
results in knock-on changes in ld-analyse, ld-process-vits and
ld-dropout-correct. Following ld-decode's usual style of using
static_cast even when an implicit conversion would do, I've added a few
missing casts and removed about an equal number of unnecessary ones.

The only remaining use of qreal is in ld-analyse, where it uses QPointF.