Extract the whole line when decoding VBI, not just the active area.
Testing with some (S-)VHS samples (e.g. Harry's Eden Project tape and
The Matrix) shows that it's possible for VITC to be well outside the
timings in the specification, extending outside the active area in both
directions. So make the whole line available, and search from
colourBurstEnd to the latest possible start that can still accommodate
90 bits of VITC.

As the code to find the next edge is now used three times in the VITC
decoder, make it a helper function.

The other decoders still use activeVideoStart as a start position as
before. We may want to revisit this, since sometimes it's useful to
adjust the active area and it shouldn't affect VBI decoding...