Use std::vector rather than QVector in ld-process-efm.
The way that the vectors are used here is pretty much anti-optimal for
QVector's "detach" behaviour -- whenever the code did something like
"buffer[i].isMissing()", it would result in the whole buffer being
copied! Which was enough to quickly run out of address space on a 32-bit
machine when decoding even a very short sample...

Switching to std::vector makes it a lot clearer where data is actually
being copied, and allows us to avoid some unnecessary copies in a couple
of places.