Try to correct CRC errors in the Q channel.
Since the error rate needs to be fairly low in order to make it worth
decoding EFM at all, the most likely type of error here is a single bit
being flipped. So when a CRC error is detected, try flipping each bit in
turn and checking the CRC again.

The improvement from this is very small because ld-process-efm already
guesses the timestamp when the CRC fails based on the previous frame;
it'll still fall back to this behaviour if correction doesn't work.