Fix type mismatches in qFatal arguments
These are mostly uses of .size() which returns qsizetype, a typedef for
ptrdiff_t; GCC correctly warns that %d is the wrong size for this.
There isn't a specifier that's guaranteed to work on both versions of Qt
and all platforms, so static_cast to int and use %d.

A better fix for this would be to use qFatal() << arg formatting, but
that's not available in Qt 5. (I've got a patch to convert them which
can land once we require Qt 6.)